+++ app/src/androidTest/java/com/hotmail/or_dvir/sabinesList/MyTestApplication.kt
@@ -1,0 +1,6 @@
+package com.hotmail.or_dvir.sabinesList
+
+import dagger.hilt.android.testing.CustomTestApplication
+
+@CustomTestApplication(BaseApplication::class)
+interface MyTestApplication